[idv] Consolidate retry and register states of passkey machine (#8724)
## Describe your changes
The retry state had pretty much the same functionality of the register state - it just also had the option to skip passkey registration altogether.

In a future PR, we're going to start sending context on passkey registration attempts to the backend - i didn't want to pass this context to a parent state machine. We probably didn't need the passkey registration to be a state machine initially anyways - so this PR consolidates the register + retry states into one.

Another added benefit: we've had some bugs in the past where the spacing is implemented differently across these screens - this consolidation helps to keep the UI consistent!

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Successful registration on first try:
https://github.com/onefootprint/monorepo/assets/21066341/26008ac9-52a7-4987-b9d9-07164b86dc7f

Failure a few times and then successful registration:

https://github.com/onefootprint/monorepo/assets/21066341/289de348-0942-44a1-9d31-5c299c2de34f

Failure and then skip:

https://github.com/onefootprint/monorepo/assets/21066341/06391954-614f-411b-a832-ac36619b88fa