[FP-7365][dashboard] Open modal to copy/send link from user timeline event (#8696)
## Describe your changesr a
We've long had the ability to generate a new link to hosted bifrost from the timeline event for a "workflow triggered" event. This button used to be called "Copy link," which was always confusing since we actually generated a new link.

And, i recently discovered that [this button didn't even work in safari](https://onefootprint.slack.com/archives/C05SN1UTZLY/p1711047992007049). So, I clarified this experience entirely by updating the button to say "Generate new link," and have this bring you back to the familiar screen where you can copy/send a link.

## Caveats
There's a lot of logic to control the behavior in this modal to copy/send the link via SMS, so i wanted to share the code between this timeline event and the "Request more info" action where it normally appears. It's made complicated by the fact that the "Request more info" feature shows this modal as the second of two screens on a single dialog. So I ended up using a hook-like paradigm to return the dialog contents + dialog properties. Let me know if you have other ideas on how to build this

## Test Plan
Locally:

I artificially increased the loading time to show the loading state:
https://github.com/onefootprint/monorepo/assets/21066341/abca08fa-7cff-4c28-875b-b6faa39f45b9

And with an error:

https://github.com/onefootprint/monorepo/assets/21066341/8e28af40-da10-4c74-beb1-e92538abfa2f