[idv] Support multiple consecutive loading screens in requirements machine (#8788)
## Describe your changes
We've long had a problem where two adjacent loading screens cause the
loading indicator to "jump around" or "skip" back to the starting point,
because we're making two different instances of the loading indicator
for each different screen. To get around this, we've done some movement
of code to try to group adjacent network requests into a single state
machine state. But unfortunately, this requires us to occasionally
duplicate code, solely to display the loading indicator properly.

This takes a new approach: states that do nothing but load will render
nothing. And then, the state machine will have one loading indicator
that it renders if the current state is just a loading screen. So, we
end up keeping the invariant that only one loading indicator is
instantiated even as the state machine's state transitions

In this PR, this also allows us to simplify the checkRequirements state
drastically - it used to (1) optionally initialize the onboarding (2)
check for requirements and (3) optionally trigger the process
requirement if it was the first req, and then refetch requirements. (3)
was duplicate code that we could delete, and (1) is now a separate state
that is solely responsible for initializing the onboarding

## Caveats
This only solves the adjacent loading problem for the onboarding
requirements machine. We could take a similar approach across other
machines too, but it will be more invasive

## Test Plan
I artificially added some sleeps on the backend to each of the endpoints
that require a loading screen, and then printed the state machine's
current state. This video shows that as we transition between states in
the machine, the loading indicator does not jump around


https://github.com/onefootprint/monorepo/assets/21066341/fa46395a-50a3-48fc-9e4e-9db37b0b0fa0