[Neuroid]: Identify email input (#8740)
## Describe your changes

We're about to integrate footprint with neuroid
(https://www.neuro-id.com/). This platform will allow us to detect
things like behavioral frauds.

To integrate neuroid, we need to identify our inputs with a unique data
attribute. I'm calling it now as `data-name` (we could use `name`
directly, but I don't want to create this dependency on this specific
attribute as it could be important for us in the future).

The idea is to identify all the occurrences - for now, I'm excluding
KYB.

This is just the first PR, but I need to add all the inputs. The
`data-name` could be easily changed to something else of course, and we
could move these names to a constant (maybe to `core`, given we also
need for ob-components)

Suggestions are welcome!

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->