[api] Add endpoint to create a new auth token with reduced scopes (#8810)
## Describe your changes
We'll start to use this in the components SDK to generate a new auth
token that has limited scopes to only vault data. This auth token will
be the only one passed back to the components SDK, so the tenant doesn't
have unfettered access to register auth credentials, decrypt data, etc

## Caveats
Today, vaulting an email also has a sneaky ability to register an
authentication credential since we support logging in with an unverified
email address. I still need to address this + further separate these in
a future PR

## Test Plan
Integration test