[footprint-js] Add support for invoking the components SDK (#8801)
## Describe your changes
Adds some internal mechanisms to footprint-js that are only intended to be used by the components SDK

We introduce a new `kind` to the `footprint.init()` invocation, "components." this ends up using id.onefootprint.com under the hood and `verify_v1` sdk args. BUT, it has a few differences:
- can only be invoked with an ob public key, no auth token
- supports the `onHandoffToComponentsSdk` callback (name pending) for components SDK to get an auth token to continue vaulting data

## Caveats

We don't really want customers to know that these events/functionalities exist outside of the components SDK. Maybe in the future we'll move the components SDK into footprint-js so it can use some internal methods that aren't exposed publicly. Or maybe one day we'll make a `footprint-js-internal` that is wrapped by `footprint-js` and the components SDK

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->