[bifrost] Propagate new sdk argument isComponentsSdk throughout idv machines (#8807)
## Describe your changes
Super boring change, just starting to read a new SDK arg and propagate it through our web of state machines so we can use it in a follow-up PR

This flag is kind of like `isTransfer` - idv will run either in an `isComponentsSdk` context or not, and `isComponentsSdk` will cause a few branches in behavior in bifrost.

## Caveats
Lots of boilerplate to pass through. Maybe better to assemble a context object inside bifrost and pass the whole thing around ¯\_(ツ)_/¯

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->