Propagate isInIframe through idv machines (#8827)
## Describe your changes
We have a lot of behavior that changes whether or not bifrost is running inside of an iframe. We probably want to write some unit tests for this, and it is currently pretty hard to do so by mocking the `checkIsInIframe` function. Instead, let's just pass this as a context variable throughout the state machines

## Caveats
Again, kind of clunky how many places need to be updated for a new piece of global state... i might just want to make a single IdvContext object that is passed to every sub-machine

## Test Plan

Unit test in next PR

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->