[bifrost] Alter bifrost behavior when running for components SDK (#8799)
## Describe your changes
We'll soon support an invocation of footprint-js specific for our components SDK. This `isComponentsSdk` flag will be passed to bifrost. In a previous PR, we propagated this flag through the state machines so it's generally available. When `isComponentsSdk` is true, bifrost should behave a little bit differently when running specifically for the components SDK:

- we won't show the "confirm" screen since the user will have just finished entering their data into the components SDK
- after calling `POST /hosted/onboarding` and before rendering any requirements in the requirements machine, we will instead "relay" to the components SDK. we will send an auth token (that should be scoped to only do the operations necessary) to the components SDK that it can use to vault data. we will then wait until we receive a postmessage from the components SDK telling us it is done vaulting data, and we will continue to handle remaining requirements
- (unimplemented) we may want to hide the "log in with different account" option from the identify machine?

Context on discussion for this architecture: https://onefootprint.slack.com/archives/C06QKF0DSF4/p1711414047271419

## Caveats

[x] TODO do not merge until we create a scoped token that can only vault data?

## Test Plan
Locally, some integration tests