[demos] Update the onboarding components demo to use new footprint-js (#8802)
## Describe your changes
In the PRs below this, we've set up bifrost to operate in a special mode
for the components SDK: after the user logs in, bifrost will "relay" an
auth token with limited scope back to the components SDK, which it can
use to vault data. When the components SDK is done vaulting data, it
calls `fp.handoff()` tl relay back to bifrost to finish handling any
requirements, including liveness, ID doc, process, and more.

## Caveats
I realize one-click onboardings will show the authorize screen too.
maybe that's desired? not sure


## Test Plan
Locally:


https://github.com/onefootprint/monorepo/assets/21066341/ff3b2bbf-557f-4126-8ba8-909b2ad53168