Embed the UserSession struct inside of our loaded auth context (#8875)
## Describe your changes
Just a little bit of cleanup to make the next PRs easier.

We've long done this strange thing where we unpack all of the properties of the UserSession onto the UserSessionContext struct after loading the UserSession from the database. We strangely had implemented the utils for making a new UserSession on the UserSessionContext struct.

This just prevents us from doing unnecessary packing, and makes it easier to tell which fields on UserSession are used, and makes it easier to add new fields to UserSession

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->