Include entire history of purposes for creating auth token in session (#8877)
## Describe your changes
We added this `purpose` field to the UserSession a while back to encode some metadata on the reason why an auth token was created. But, our auth model often allows creating one auth token using the permissions of another auth token, or stepping up a token that has no permissions into one that does have permissions. So, we'd often lose some history of the reasons why scopes were added to sessions over time.

This broadens the use cases covered by the TokenCreationPurpose enum and starts to track the "purpose" each time a token is created _or_ updated.

This will particularly come in handy as we introduced our first case where we issue a new auth token that has _reduced_ scopes for the components SDK. We want to make sure that this token issued for the components SDK can never step itself up to a token that has additional scopes.

## Caveats

Note: there's only really one place where we mutate a token in place, and that's in `POST /hosted/onboarding`. All other use cases issue a brand new token with new scopes, so knowledge of a previous token with no permissions (like in the links we allow dashboard users to send to their customers) does not necessarily give you access to a new token with permissions.

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->