Fix the serde rename on UserAuthScope (#8879)
## Describe your changes
We're making some other breaking changes to UserSession, so it's a good time to batch in these breaking changes too :)

We'll start serializing the scopes as snake_case, but support the old CamelCase serializations. When all the legacy user sessions have expired, we'll drop the serde aliases and consolidate to only use UserAuthScope (and get rid of UserAuthGuard)

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Unit test for serialization

Did a test locally where I created a token on the old server version, then updated to run the server with this code. I was still able to use the auth token created by the old server version with the new server version