O.C (2): Fix user data (#8848)
### TL;DR

This PR updates the `Provider` to use the `userData` prop to set form
default values.

### What changed?

In the `Provider` component, we are now using the `userData` prop to set
default values for the fields in the form. This means that if a user's
information is included in the `userData` prop, that information will be
auto-filled when they encounter the form.

### How to test?

Test this by passing a `userData` prop to the `Provider` with user
information and ensure that the form fields fill in correctly.

### Why make this change?

This change allows for a more streamlined form filling process for
users, enhancing UX.

---

## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->