Create general alert on requirements machine stuck (#8922)
## Describe your changes
In light of a lack of end-to-end flow alerting, this is a creative way to tell if users are getting stuck, covering a whole category of bugs.

We added a similar alert a few weeks back to alert when a user entered the collect KYC data machine multiple times - this now generalizes that alert to tell us when the user is stuck on _any_ requirement.

This exploits the fact that before handling the next requirement,  we re-fetch requirements from the backend. If the highest priority requirement from the backend is the one that the frontend thinks it just solved, we alert that the user is stuck

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally - there's one bug (solved in #8921) in which users end up stuck. We logged properly in this case

https://github.com/onefootprint/monorepo/assets/21066341/3f349e71-9c0a-420c-ab0c-58abdd722588

Also tested that the normal flow doesn't cause alerts locally