[idv] Support multiple id doc requirements in transfer machine (#8924)
## Describe your changes
When we need to upload multiple documents, we right now just show copy based on the first document requirement. This now passes all document requirements to the transfer machine so it can display vague copy to upload "documents" when there are multiple.

I believe today the only use case of this is proof of SSN, which requires both an ID doc and ssn card

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Locally:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/9b0fa150-b065-44eb-b872-cd3ee896bb44.png)