Deprecate old fields on UserSession (#8881)
## Describe your changes
Won't merge this until the PR below this has been deployed for 3 days, the maximum duration for which a user token can live.

This makes a few breaking changes to the UserSession struct - we are combining them into one PR because they all need the same 3-day period to bake before they are safe to merge:
- Only accept snake case serialization of UserAuthScope. We changed this serialization from camelcase to snake case in #8879
- Start reading `purposes` instead of `purpose`, and stop reading / writing `purpose`. We migrated to `purposes` in #8877
- Stop deserializing `is_implied_auth`. we've been able to rm this one for quite some time - we store the vec of AssociatedAuthMethods in the auth session, which contains information on whether each auth method was implicit or explicit.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Made a token with the server version below this. Was able to use the token with this server version too, showing we have been writing everything we need for 3d in order to safely merge this change