[FP-7398] Only accept VaultData permissions in PATCH /hosted/user/vault API (#8883)
## Describe your changes
In #8810, we started giving all tokens the `UserAuthScope::VaultData` scope who needed to vault data. I'll merge this PR with the ones below this, probably on monday 4/1, which will be plenty of time for all of the old auth tokens to have expired. In other words, by the time this is merged, every token that has the `SignUp` scope will also have the `VaultData` scope.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->