Fingerprint dupes (#8760)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Adds `get_dupes` method for fetching dupes for a given SV, to be used by
an API the FE calls to populate our new dupes table.
Adds query on `fingerprint` to query for ssn/email/phone dupes.

For a given kind of data:

- If there are duplicates from SV's on the same tenant (ie `tenant`
scoped fingerprint's), then we'll return the `fp_id`'s for those dupes.
(And in future would add other things like current status, onboarding
time, etc for populating other aspects of the dupes table).
- If there are duplicate from vault(s) on different tenants (ie `global`
scoped fingerprint's), then we just return a count of Vault's that are
dupes. (a Vault that is being returned as a same-tenant dupe wouldn't
factor into this count)

## Caveats

<!--- Are there any caveats to this changeset? -->
existing indexes/query plan seemed reasonable but still need to analyze
performance of the query a bit more in depth

## Test Plan

<!--- How did you test your changes? -->

integration tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->