Store when phone/email is bootstrapped into identify (#9011)
## Describe your changes
Data may be bootstrapped into the identify flow, but it's also possible that bootstrapped data gets overwritten by the user. Maybe the phone is invalid, or maybe they just choose to log in with a different account entirely.

So, this allows the identify machine to start properly tracking the source of the data - bootstrap data has an `isBootstrap` flag set to true, and any time the data is updated, we wipe that flag and set it to false.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->