Fix flake in partner tenant integration test fixtures (#9037)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
Example failure: https://github.com/onefootprint/monorepo/actions/runs/8573096040/job/23497107273?pr=9036

Providing a fixed ID to a get-or-create operation on the `partner_tenant` table with a unique constraint on the ID is a race condition that can lead to a flake from constraint violation errors.

Would be nice to fix this in a cleaner way than adding retries, but doing the same thing we do for the tenant fixtures for now.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->