Multi edit rule backtesting (#9020)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Modify backtesting endpoint to support taking in a bulk set of multiple
deletions/edits/additions. Retrieves the current set of rules, applies
those hypothetical changes, and then evaluates that new set of rules
against the sample of onboardings and returns those results and some
summary stats.

## Caveats

<!--- Are there any caveats to this changeset? -->

In the ideal case, you are backtesting only on onboardings that have
occurred since the last rule edit was made. If that is the case then the
results of the backtest are purely a function of the hypothetical edits.

As discussed previously, there is nuance when you are backtesting over
onboardings that span other subsequent rule edits. There are basically 2
kinds of errors this introduces:

- (1) Backtest Results including diff that _does not_ come from the
Hypothetical Edits
- Some onboarding’s backtest result is a function of not only the
Hypothetical Edits but also every other rule edit that occurred since
the onboarding and now.
- This even means that you could have 0 changes in your Hypothetical
Edits (ie just backtesting the current rule set), and you would see diff
in the backtest results.
  - Why does this matter?
- The user is expecting to gauge what impact their Hypothetical Edits
will have but now what they are seeing is some function of other
factors. Furthermore, it is not easy to know how much impact those other
factors are having.
- (2) Backtest Results _not_ including diff from the Hypothetical Edits
- eg: User onboards, then Rule A is added, and now we are backtesting
Deleting Rule A.
- There is no chance for any diff to come from this user in the
backtest. However, Rule A could possibly actually trigger for the user
and deleting it would indeed have an impact for that user.
- eg: Rule A_v1 added, user onboards, now rule edited to Rule A_v2, and
now we are backtesting another edit Rule A_v3.
- If Rule A_v1 and Rule A_v3 evaluate to the same thing, but this is
different from what Rule A_v2 evaluates to, then no diff will be shown
in the backtest.
  - Why does this matter?
- The user is expecting to gauge what impact their Hypothetical Edits
will have, and there could be changes that are actually very impactful,
but now there is some population of onboardings that are “impervious” to
these changes just by virtue of the rule not existing or existing in a
different from at the time of their onboarding. This can make the edits
appear less impactful in the backtest than they actually are.

## Test Plan

<!--- How did you test your changes? -->

not a whole lot, def need more testing here

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->