Fix 500 error in PATCH /org (#9025)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
This isn't a unique problem to `PATCH /org`, but I'm fixing this one to demonstrate the problem. When there are no updates in a Diesel update struct (i.e. all `None`), it yields an error: `ApplicationError(ApiError(Database(DbError(QueryBuilderError("There are no changes to save. This query cannot be built")))))`.

Prior to https://github.com/onefootprint/monorepo/pull/9024 this would have panicked, but now it's a 500 error.

Many of our PATCH queries probably have this problem, but at least a well-behaving frontend would usually not send empty PATCH requests.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->