Add GET /partner/auth/roles API (#9036)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
Currently we're using /org/auth/login for partner login, but the support for partner tenants is only partial. For example, logout is not supported and if the user is part of multiple partner orgs, they can't log in to any since they can only get a `WorkOsSession`. In this PR I begin the process of separating out partner auth methods. I'll use the `org_common/authn` package to share common auth logic.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->