Only compute collect KYC transitions based on missing data (#9053)
## Describe your changes
In #9051, we'll again start properly passing when email/phone are
bootstrapped or not to the KYC data machine (before, we would always say
they were "bootstrapped" to the KYC data machine). This had a bug - we
would incorrectly accidentally show the email screen in collect KYC
data.



https://github.com/onefootprint/monorepo/assets/21066341/98d9b050-cfd3-412a-9dd9-6444dc5812b4



#9052 was my first attempt to fix that bug, but I realized it's just
adding another difficult-to-understand setting to the machine in order
to work around the transitions that are kind of difficult to understand.
I think this PR is a better approach - if the backend reports that a CDO
isn't missing, the frontend shouldn't even show the screen.

I cannot think of a case in which we'd want to show the data collection
screen on the frontend when the backend says everything on that screen
is already collected. So I think this was just a pre-existing logic bug
that was almost never evoked. This [existing
logic](https://github.com/onefootprint/monorepo/blob/46f7ef158ff4566013fe412e688c39bee2fea99d/frontend/packages/idv/src/plugins/collect-kyc-data/components/navigation-header/utils/current-step-from-missing-attributes.ts#L19-L22)
already did something similar and only paid attention to potentially
missing attributes.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan



https://github.com/onefootprint/monorepo/assets/21066341/23030767-9edc-4d34-803b-bb99173e6ebd