Save when contact info is added via bifrost in signup challenge (#8970)
## Describe your changes
Updates the signup challenge API to support specifying the raw phone number alongside a boolean indicating whether the phone number was bootstrapped into the flow.

In the future, we may limit whether users can use the "login with unverified email" feature when the email was bootstrapped: https://onefootprint.slack.com/archives/C06QKF0DSF4/p1712099924817119

## Caveats

is_bootstrap is self-reported by the client. it makes me wary of using it for any real business logic on the backend since someone with a user auth token could spoof it. the good news is the code that reports is_bootstrap runs inside of our iframe, so in theory, it cannot be mucked with.

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->