feat(compliance): FP-7360 doc viewer route (#9056)
### Warning
The CI task will stop failing once the two other dependencies are
merged.

- **#9056** 👈
- **#9054**
- **#9055**

## Describe your changes
This PR introduces the doc view route of `apps/partners`. This route is
responsible to view, review and approve or reject documents

The **doc** route takes 3 parameters
`doc/[partnerId]/[documentId]/[submissionId]`. These parameters are
required to submit a review to the backend

### Screenshots


![image](https://github.com/onefootprint/monorepo/assets/5114769/263dcec4-a6ae-4ee0-9d8c-4af1f8cf7861)


![image](https://github.com/onefootprint/monorepo/assets/5114769/6f422151-b95d-4f47-b683-35ecb79b49b4)