Better track when data is dirty in collect KYC machine (#9074)
## Describe your changes
The tracking of `dirty` on the KYC machine is pretty new. In the past,
we treated dirty data and bootstrap data pretty similarly
([ex](https://github.com/onefootprint/monorepo/blob/5d760beca1fedf60830045a2a4dc56970b1b90ee/frontend/packages/idv/src/plugins/collect-kyc-data/hooks/use-sync-data/utils/get-request-data/get-request-data.ts#L32)),
and even passed data into the KYC machine as "bootstrap" (until #9014)
when it needed to be sent to the backend.

Instead, we now properly track when the KYC machine is initialized with
"dirty" data that needs to be sent to the backend.

## Caveats

I even think after some more cleanup we could get rid of the linked
codepath above that treats bootstrap data automatically as "dirty"

NOTE: we still need to update all of the use-convert-form-datas to
properly preserve when a piece of data is "dirty," even if its value
hasn't changed... this code is pretty involved, so i will update in a
follow up PR

## Test Plan
Caught this because #9014 broke a very little-used codepath where (1) we
log into an existing user with no email and (2) the identify machine
collected an email and passes it into the KYC machine. When e2e tests
are more stable, I'll write an e2e test for this niche case.

Before this PR:



https://github.com/onefootprint/monorepo/assets/21066341/72d5ac56-b6e7-4b45-b051-b4374aaaf40b


After this PR:



https://github.com/onefootprint/monorepo/assets/21066341/dd812348-3cfd-4c0c-88f4-412fa534537c

(NOTE: in a follow-up PR, we're going to stop even showing the email
page when it's already collected)